+++ src/constraints.jl
@@ -58,11 +58,11 @@ function constraint_gate_target_condition(qcm::QuantumCircuitModel)
     
     # For correct implementation of this, use MutableArithmetics.jl >= v0.2.11
     if decomposition_type == "exact"
-        JuMP.@constraint(qcm.model, sum(qcm.variables[:V_var][:,:,n,depth] * qcm.data["gates_real"][:,:,n] for n=1:num_gates) .== qcm.data["target_gate"])  
+        JuMP.@constraint(qcm.model, sum(qcm.variables[:V_var][:,:,n,depth] * qcm.data["gates_real"][:,:,n] for n=1:num_gates) .== qcm.data["target_gate"])
     
     elseif decomposition_type == "approximate"
         JuMP.@constraint(qcm.model, sum(qcm.variables[:V_var][:,:,n,depth] * qcm.data["gates_real"][:,:,n] for n=1:num_gates) .== qcm.data["target_gate"][:,:] + qcm.variables[:slack_var][:,:])  
-        
+
     end
 
     return