+++ CHANGELOG.md
@@ -4,6 +4,7 @@ QuantumCircuitOpt.jl Change Log
 
 ### v0.1.9
 - Updated toffoli input circuit to make it feasible for the commuting gate constraints 
+- Bug fix in the qubit specificity of RY gate in 3 qubits. 
 
 ### v0.1.8
 - Added support for specificity of qubits in U3 and Rotation (RX, RY, RZ) gates (Major change). Example: U3_1, RX_2, RZ_3, etc